user_service/requirements.txt
fastapi==0.115.7
passlib==1.7.4
pydantic==2.10.6
PyJWT==2.10.1
SQLAlchemy==2.0.37
uvicorn==0.34.0
